requirements.txt
pandas>=1.5
faiss-cpu>=1.7
flask>=2.0
flask-cors>=3.0
gunicorn
openpyxl
google-generativeai
langchain
langchain-google-genai